Move AbinsFrequencyPowderGeneratorTest from algorithms to scripts
Remove superfluous test file in algorithms

The existing test in scripts covers the function fairly well, but the
new part checks the filtering of high frequencies is functioning as
intended.
